requirements.txt
github-flask
peewee
# flask-peewee
-e git+https://github.com/coleifer/flask-peewee.git#egg=flask-peewee
-e git+git@github.com:hapyak/flask-peewee-swagger.git#egg=flask-peewee-swagger
